requirements.txt
datasets==4.1.1
huggingface-hub==0.34.4
keras==3.11.3
matplotlib==3.10.6
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.0
peft==0.17.1
pillow==11.3.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tf_keras==2.20.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@6eb32558424c6a461252769f7aca3b3de834912b
triton==3.3.1
trl==0.23.1


